Zomato/architecture/phase_4_llm_recommendation/requirements.txt
groq>=0.9.0
flask>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
